Refactor file structure
Move all functions for replacing vowels into their own python package.
